+++ neptune/new/internal/backends/neptune_backend.py
@@ -29,6 +29,7 @@ from neptune.new.internal.backends.api_model import (
     FloatSeriesAttribute,
     IntAttribute,
     Project,
+    Workspace,
     StringAttribute,
     StringSeriesAttribute,
     StringSetAttribute,
@@ -53,6 +54,17 @@ class NeptuneBackend:
 
     @abc.abstractmethod
     def get_project(self, project_id: str) -> Project:
+        pass
+
+    @abc.abstractmethod
+    def get_available_projects(self,
+                               workspace_id: Optional[str] = None,
+                               search_term: Optional[str] = None
+                               ) -> List[Project]:
+        pass
+
+    @abc.abstractmethod
+    def get_available_workspaces(self) -> List[Workspace]:
         pass
 
     @abc.abstractmethod